[not verified] move title handling to uploader progress cmp
