+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/videopress-uploader/style.scss
@@ -63,6 +63,7 @@
 
 	&__title {
 		margin-bottom: 15px;
+		width: 100%;
 	}
 
 	&__upload-link {
